-fix folder structure -add new package json -edit main to correct ts-config path
