+++ apps/browser/src/autofill/content/components/lit-stories/buttons/edit-button.lit-stories.ts
@@ -2,17 +2,16 @@ import { Meta, StoryObj } from "@storybook/web-components";
 
 import { Theme, ThemeTypes } from "@bitwarden/common/platform/enums/theme-type.enum";
 
-import { ActionButton } from "../../../buttons/action-button";
+import { EditButton } from "../../buttons/edit-button";
 
 type Args = {
-  buttonText: string;
-  disabled: boolean;
-  theme: Theme;
   buttonAction: (e: Event) => void;
+  buttonText: string;
+  disabled?: boolean;
+  theme: Theme;
 };
-
 export default {
-  title: "Components/Buttons/Action Button",
+  title: "Components/Buttons/Edit Button",
   argTypes: {
     buttonText: { control: "text" },
     disabled: { control: "boolean" },
@@ -27,7 +26,7 @@ export default {
   },
 } as Meta<Args>;
 
-const Template = (args: Args) => ActionButton({ ...args });
+const Template = (args: Args) => EditButton({ ...args });
 
 export const Default: StoryObj<Args> = {
   render: Template,
