+++ apps/browser/src/autofill/content/components/lit-stories/icons/icons.lit-stories.ts
