+++ apps/browser/src/autofill/content/components/lit-stories/rows/button-row.lit-stories.ts
@@ -2,27 +2,23 @@ import { Meta, StoryObj } from "@storybook/web-components";
 
 import { Theme, ThemeTypes } from "@bitwarden/common/platform/enums/theme-type.enum";
 
-import { CloseButton } from "../../../buttons/close-button";
+import { ButtonRow } from "../../rows/button-row";
 
 type Args = {
-  handleCloseNotification: (e: Event) => void;
   theme: Theme;
 };
+
 export default {
-  title: "Components/Buttons/Close Button",
+  title: "Components/Rows/Button Row",
   argTypes: {
     theme: { control: "select", options: [...Object.values(ThemeTypes)] },
-    handleCloseNotification: { control: false },
   },
   args: {
     theme: ThemeTypes.Light,
-    handleCloseNotification: () => {
-      alert("Close button clicked!");
-    },
   },
 } as Meta<Args>;
 
-const Template = (args: Args) => CloseButton({ ...args });
+const Template = (args: Args) => ButtonRow({ ...args });
 
 export const Default: StoryObj<Args> = {
   render: Template,
